requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
twilio==9.3.3
apscheduler==3.10.4
python-multipart==0.0.12
